+++ website/src/components/demos/alert.demo.tsx
@@ -1,24 +1,23 @@
 import { InfoIcon } from 'lucide-react'
-import { Stack } from 'styled-system/jsx'
-import { Alert, type AlertProps } from '~/components/ui/alert'
-import { Heading } from '~/components/ui/heading'
-import { Icon } from '~/components/ui/icon'
-import { Text } from '~/components/ui/text'
+import {
+  Alert,
+  AlertContent,
+  AlertDescription,
+  AlertIcon,
+  AlertTitle,
+  type AlertProps,
+} from '~/components/ui/alert'
 
 export const Demo = (props: AlertProps) => {
   return (
     <Alert {...props}>
-      <Stack gap="4" direction={{ base: 'column', sm: 'row' }}>
-        <Icon color="fg.emphasized">
-          <InfoIcon />
-        </Icon>
-        <Stack gap="1">
-          <Heading as="h5" fontWeight="medium">
-            Browser update available
-          </Heading>
-          <Text color="fg.muted">For the best experience, please update your browser.</Text>
-        </Stack>
-      </Stack>
+      <AlertIcon asChild>
+        <InfoIcon />
+      </AlertIcon>
+      <AlertContent>
+        <AlertTitle>Browser Update available</AlertTitle>
+        <AlertDescription>For the best experience, please update your browser.</AlertDescription>
+      </AlertContent>
     </Alert>
   )
 }