+++ website/src/components/ui/text.tsx
@@ -1,16 +1,14 @@
-import type { PropsWithChildren } from 'react'
 import { styled, type HTMLStyledProps } from 'styled-system/jsx'
 
 type As = 'p' | 'span' | 'div' | 'label'
 
-type TypographyProps = PropsWithChildren<{
+export type TextProps = {
   as?: As
-}> &
-  HTMLStyledProps<As>
+} & HTMLStyledProps<As>
 
-export const Text = (props: TypographyProps) => {
+export const Text = (props: TextProps) => {
   const { as = 'p', ...rest } = props
-
   const Component = styled(as)
+
   return <Component {...rest} />
 }