+++ packages/presets/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - Added support for icons in the `Badge` component.
 
+### Changed
+
+- Revised `Alert` component.
+
 ## [0.16.0] - 2023-09-27
 
 ### Added